Rename files (from `together` to `talk`) - JavaScript/TypeScript files together => braveTalk - Rename extension API - Rename components/brave_together => components/brave_talk - Rename file `togetherIcon` => `braveTalkIcon` - Rename file `brave-together-icon.tsx` => `brave-talk-svg.tsx`
