+++ browser/extensions/BUILD.gn
@@ -1,7 +1,7 @@
 import("//brave/browser/ethereum_remote_client/buildflags/buildflags.gni")
 import("//brave/components/binance/browser/buildflags/buildflags.gni")
 import("//brave/components/brave_rewards/browser/buildflags/buildflags.gni")
-import("//brave/components/brave_together/buildflags/buildflags.gni")
+import("//brave/components/brave_talk/buildflags/buildflags.gni")
 import("//brave/components/brave_wallet/browser/config.gni")
 import("//brave/components/brave_wayback_machine/buildflags/buildflags.gni")
 import("//brave/components/brave_webtorrent/browser/buildflags/buildflags.gni")
@@ -195,12 +195,12 @@ source_set("extensions") {
 
   if (brave_together_enabled) {
     sources += [
-      "api/brave_together_api.cc",
-      "api/brave_together_api.h",
+      "api/brave_talk_api.cc",
+      "api/brave_talk_api.h",
     ]
 
     deps += [
-      "//brave/components/brave_together/browser",
+      "//brave/components/brave_talk/browser",
       "//brave/components/ntp_widget_utils/browser",
     ]
   }
