+++ browser/extensions/api/brave_talk_api.cc
@@ -3,13 +3,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/browser/extensions/api/brave_together_api.h"
+#include "brave/browser/extensions/api/brave_talk_api.h"
 
 #include <string>
 #include <memory>
 
 #include "base/environment.h"
-#include "brave/components/brave_together/browser/regions.h"
+#include "brave/components/brave_talk/browser/regions.h"
 #include "brave/components/ntp_widget_utils/browser/ntp_widget_utils_region.h"
 #include "chrome/browser/profiles/profile.h"
