+++ browser/sources.gni
@@ -32,7 +32,7 @@ import(
 import("//brave/components/brave_referrals/buildflags/buildflags.gni")
 import("//brave/components/brave_rewards/browser/buildflags/buildflags.gni")
 import("//brave/components/brave_sync/buildflags/buildflags.gni")
-import("//brave/components/brave_together/buildflags/buildflags.gni")
+import("//brave/components/brave_talk/buildflags/buildflags.gni")
 import("//brave/components/brave_wallet/common/buildflags/buildflags.gni")
 import("//brave/components/brave_wayback_machine/buildflags/buildflags.gni")
 import("//brave/components/brave_webtorrent/browser/buildflags/buildflags.gni")
@@ -250,7 +250,7 @@ if (enable_sidebar) {
 }
 
 if (brave_together_enabled) {
-  brave_chrome_browser_deps += [ "//brave/components/brave_together/browser" ]
+  brave_chrome_browser_deps += [ "//brave/components/brave_talk/browser" ]
 }
 
 if (enable_tor) {
