+++ common/extensions/api/brave_together.json
@@ -7,7 +7,7 @@
     "namespace": "braveTogether",
     "description": "Use the <code>chrome.braveTogether</code> API to to receive info pertaining to the Brave Talk feature.",
     "compiler_options": {
-      "implemented_in": "brave/browser/extensions/api/brave_together_api.h"
+      "implemented_in": "brave/browser/extensions/api/brave_talk_api.h"
     },
     "events": [
     ],
