+++ components/brave_new_tab_ui/components/default/braveTalk/assets/brave-talk-svg.tsx
