+++ components/brave_new_tab_ui/components/default/braveTalk/style.tsx
