+++ components/brave_new_tab_ui/components/default/footer/braveTalkTooltip/braveTalkIcon.tsx
