+++ components/brave_talk/browser/BUILD.gn
@@ -1,0 +1,11 @@
+import("//brave/components/brave_talk/buildflags/buildflags.gni")
+
+source_set("browser") {
+  sources = [
+    "regions.h",
+  ]
+
+  public_deps = [
+    "//brave/components/brave_talk/buildflags",
+  ]
+}
