+++ components/brave_talk/browser/regions.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_TOGETHER_BROWSER_REGIONS_H_
-#define BRAVE_COMPONENTS_BRAVE_TOGETHER_BROWSER_REGIONS_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_TALK_BROWSER_REGIONS_H_
+#define BRAVE_COMPONENTS_BRAVE_TALK_BROWSER_REGIONS_H_
 
 #include <string>
 #include <vector>
@@ -17,4 +17,4 @@ const std::vector<std::string> unsupported_regions = {
 
 }  // namespace brave_together
 
-#endif  // BRAVE_COMPONENTS_BRAVE_TOGETHER_BROWSER_REGIONS_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_TALK_BROWSER_REGIONS_H_
