--- components/brave_together/browser/BUILD.gn
@@ -1,11 +1,0 @@
-import("//brave/components/brave_together/buildflags/buildflags.gni")
-
-source_set("browser") {
-  sources = [
-    "regions.h",
-  ]
-
-  public_deps = [
-    "//brave/components/brave_together/buildflags",
-  ]
-}
